# openapi.yaml
openapi: 3.0.3
info:
  title: E2EE PubSub Messaging API
  description: |
    End-to-end encrypted messaging system with capability-based authorization.
    
    ## Architecture
    - **Spaces**: Access control boundary (one shared symmetric key)
    - **Topics**: Message streams within a space (independent blockchain-like chains)
    - **State**: Per-space key-value store (base64-encoded data, context-dependent interpretation)
    - **Capabilities**: Granular, signed permissions using public key cryptography
    
    ## Security Model
    - Zero-knowledge server (can't decrypt message content)
    - Capability-based authorization with signature verification
    - Message integrity via hash chains
    - Space creator (space public key) has implicit admin rights
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
  /spaces/{space_id}/auth/challenge:
    post:
      tags: [Authentication]
      summary: Request authentication challenge
      description: Get a random nonce to sign for authentication
      parameters:
        - $ref: '#/components/parameters/SpaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [public_key]
              properties:
                public_key:
                  type: string
                  pattern: '^[A-Za-z0-9_-]{44}$'
                  minLength: 44
                  maxLength: 44
                  description: Typed user identifier (44-char URL-safe base64, Ed25519 public key with header byte)
      responses:
        '200':
          description: Challenge issued successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  challenge:
                    type: string
                    format: base64
                    description: Random nonce to sign
                  expires_at:
                    type: integer
                    format: int64
                    description: Unix timestamp in milliseconds when challenge expires
        '400':
          $ref: '#/components/responses/BadRequest'
  /spaces/{space_id}/auth/verify:
    post:
      tags: [Authentication]
      summary: Verify signed challenge and get JWT
      description: Submit signed challenge to authenticate and receive JWT token
      parameters:
        - $ref: '#/components/parameters/SpaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [public_key, signature, challenge]
              properties:
                public_key:
                  type: string
                  pattern: '^[A-Za-z0-9_-]{44}$'
                  minLength: 44
                  maxLength: 44
                  description: Typed user identifier (44-char URL-safe base64)
                signature:
                  type: string
                  format: base64
                  description: Signature of challenge
                challenge:
                  type: string
                  format: base64
                  description: The challenge that was signed
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT bearer token
                  expires_at:
                    type: integer
                    format: int64
                    description: Unix timestamp in milliseconds when token expires
        '401':
          $ref: '#/components/responses/Unauthorized'
  /spaces/{space_id}/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh JWT token
      description: Get a new JWT token before the current one expires
      parameters:
        - $ref: '#/components/parameters/SpaceId'
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: New JWT bearer token
                  expires_at:
                    type: integer
                    format: int64
        '401':
          $ref: '#/components/responses/Unauthorized'
  /spaces/{space_id}/state/{path}:
    get:
      tags: [State]
      summary: Get state value
      description: Retrieve state data as base64-encoded string
      parameters:
        - $ref: '#/components/parameters/SpaceId'
        - name: path
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-zA-Z0-9._-]+(/[a-zA-Z0-9._-]+)*$'
          description: |
            State path using slug-formatted segments (alphanumeric, dots, hyphens, underscores).
            Examples: "profiles/alice", "topics/general/messages", "files/photo.jpg"
            Wildcards like {self}, {any} are NOT allowed in user-created paths.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: State retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StateEntry'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    put:
      tags: [State]
      summary: Set state value
      description: |
        Create or update state data with cryptographic signature verification.
        All state entries must be signed - the server verifies the signature before storing.
      parameters:
        - $ref: '#/components/parameters/SpaceId'
        - name: path
          in: path
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data, signature, signed_by, signed_at]
              properties:
                data:
                  type: string
                  format: base64
                  description: Base64-encoded state data (interpretation is context-dependent)
                signature:
                  type: string
                  format: base64
                  description: Ed25519 signature over (space_id|path|data|signed_at) by signer (REQUIRED for all state writes)
                signed_by:
                  type: string
                  pattern: '^[A-Za-z0-9_-]{44}$'
                  minLength: 44
                  maxLength: 44
                  description: Typed user/tool identifier of signer (REQUIRED for all state writes)
                signed_at:
                  type: integer
                  format: int64
                  description: Unix timestamp in milliseconds when entry was signed (REQUIRED for all state writes)
      responses:
        '200':
          description: State updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  path:
                    type: string
                  signed_at:
                    type: integer
                    format: int64
                    description: Unix timestamp in milliseconds from the signed state entry
        '400':
          description: Invalid signature or missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      tags: [State]
      summary: Delete state value
      description: |
        Remove state data with cryptographic signature verification.
        Deletion requests must be signed - the server verifies the signature before deleting.
      parameters:
        - $ref: '#/components/parameters/SpaceId'
        - name: path
          in: path
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [signature, signed_by, signed_at]
              properties:
                signature:
                  type: string
                  format: base64
                  description: Ed25519 signature over (path + "DELETE" + signed_at) by signer (REQUIRED)
                signed_by:
                  type: string
                  pattern: '^[A-Za-z0-9_-]{44}$'
                  minLength: 44
                  maxLength: 44
                  description: Typed user/tool identifier of signer (REQUIRED)
                signed_at:
                  type: integer
                  format: int64
                  description: Unix timestamp in milliseconds when deletion was signed (REQUIRED)
      responses:
        '204':
          description: State deleted successfully
        '400':
          description: Invalid signature or missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /spaces/{space_id}/topics/{topic_id}/messages/{message_hash}:
    get:
      tags: [Messages]
      summary: Get specific message by hash from a topic
      description: Retrieve a single message from a specific topic using its content hash
      parameters:
        - $ref: '#/components/parameters/SpaceId'
        - $ref: '#/components/parameters/TopicId'
        - name: message_hash
          in: path
          required: true
          schema:
            type: string
            pattern: '^[A-Za-z0-9_-]{44}$'
            minLength: 44
            maxLength: 44
          description: Typed message identifier (SHA256 hash with header)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Message retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /spaces/{space_id}/stream:
    get:
      tags: [WebSocket]
      summary: WebSocket stream for real-time space messages
      description: |
        Establishes a WebSocket connection to receive real-time messages from all topics in the space.

        **Protocol**: WebSocket (ws:// or wss://)

        **Authentication**: Pass JWT token as query parameter `?token=<jwt>` or in the `Sec-WebSocket-Protocol` header.

        **Message Format**: JSON-encoded message objects (same schema as Message)

        **Connection**: Upgrade from HTTP to WebSocket using standard WebSocket handshake.
      parameters:
        - $ref: '#/components/parameters/SpaceId'
        - name: token
          in: query
          required: false
          schema:
            type: string
          description: JWT bearer token for authentication (alternative to Sec-WebSocket-Protocol header)
      responses:
        '101':
          description: Switching Protocols - WebSocket connection established
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      x-websocket:
        messageSchema:
          $ref: '#/components/schemas/Message'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/verify
  parameters:
    SpaceId:
      name: space_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^[A-Za-z0-9_-]{44}$'
        minLength: 44
        maxLength: 44
      description: Typed space identifier (44-char URL-safe base64, Ed25519 public key with header byte)
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'